refactor the folder stucture and Create Button Component
