+++ src/App.jsx
@@ -1,14 +1,13 @@
 import React from "react";
-import Counter from "./components/Counter";
 import { Provider } from "react-redux";
-import { store } from "./app/store";
+import { store } from "./redux/store/store";
+import { Counter } from "./components";
+
 
 function App() {
   return (
     <Provider store={store}>
-      <div>
-        <Counter />
-      </div>
+        <Counter/>
     </Provider>
   );
 }
