+++ src/components/index.js
@@ -1,0 +1,2 @@
+import Counter from "./counterApp/counter"
+export  {Counter}
