+++ src/redux/store/store.js
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import counterReducer from "../features/counter/counterSlice";
+import counterReducer from "../slice/counterSlice";
 
 export const store = configureStore({
   reducer: {
